Struct lastfm_parse_rs::lastfm_type::Request

pub struct Request<'rq, T> where
    T: RequestParams + Debug,  {
    pub base_url: &'rq str,
    pub api_key: &'rq str,
    pub method: Method,
    pub params: T,
}

Request information associated with a method and lastfm data type. Can be converted to a Url. No POST request support yet.

Fields
base_url: &'rq str api_key: &'rq str method: Method params: T
Methods

impl<'rq, T> Request<'rq, T> where
T: RequestParams + Debug,

pub fn as_url(&self) -> Url

Converts Request object to an Url, appends request parameters (GET only)