Trait lastfm_parse_rs::lastfm_type::RequestParams

pub trait RequestParams {
    fn append_to(&self, url: &mut Url);
}

(Maybe) Temporary solution: a trait for request parameter type that makes this type to know how to add itself to an url.

Required Methods

fn append_to(&self, url: &mut Url)