Trait lastfm_parse_rs::request::RequestParams

pub trait RequestParams {
    fn method(&self) -> &str;

    fn append_to(&self, url: &mut Url);

    fn needs_signature(&self) -> bool;

    fn needs_session_key(&self) -> bool;
}

Describes API method requests

Required methods

fn method(&self) -> &str

Returns API method name

fn append_to(&self, url: &mut Url)

Appends method parameters to given url. Please note that common and special parameters like method name, api key, session key and signature are appended automatically somewhere else

fn needs_signature(&self) -> bool

Most write (see https://www.last.fm/api/rest) API methods (except auth) need to be signed. This function indicates whether given method is one of these.

fn needs_session_key(&self) -> bool

All write (see https://www.last.fm/api/rest) API methods require user to be authenticated. For them session key has to be additionally provided in request body.

Loading content...