Trait lastfm_parse_rs::lastfm_type::LastfmType

pub trait LastfmType<'de> where
    Self: Sized + Deserialize<'de>, {
    type Outer: Deserialize<'de> + Debug + Into<Self>;
}

All lastfm json data types are wrapped with an extra object. Like this: { tag: { /* actual fields */ } } While Rust object structure has to match json's in order for serde to work, this extra wrapping is not very convenient from users perspective. This trait indicates that type is wrapped and provides deserializable wrapper struct type along with inner type conversion. This trait is used internally to automatically generate wrapper objects and hide them from user (see lastfm_t! macro).

Associated Types

type Outer: Deserialize<'de> + Debug + Into<Self>

Loading content...